requirements.txt
cfgv==3.4.0
coverage==7.6.10
distlib==0.3.9
filelock==3.16.1
flake8==7.1.1
flake8-docstrings==1.7.0
identify==2.6.4
iniconfig==2.0.0
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
requests==2.32.3
pydantic==2.10.4
PyYAML==6.0.2
snowballstemmer==2.2.0
virtualenv==20.28.0
watchdog==6.0.0
